Remove TchapRoomSummaryItem
Use the RoomSummaryItem from Element instead.
Kept the custom tchap layout and updated the item to work with it.
