+++ vector/src/main/java/im/vector/app/features/home/room/list/RoomSummaryListController.kt
@@ -16,11 +16,10 @@
 
 package im.vector.app.features.home.room.list
 
-import fr.gouv.tchap.features.home.room.list.TchapRoomSummaryItemFactory
 import org.matrix.android.sdk.api.session.room.model.RoomSummary
 
 class RoomSummaryListController(
-        private val roomSummaryItemFactory: TchapRoomSummaryItemFactory
+        private val roomSummaryItemFactory: RoomSummaryItemFactory
 ) : CollapsableTypedEpoxyController<List<RoomSummary>>() {
 
     var listener: RoomListListener? = null
